requirements.txt
alembic==0.8.9
bcrypt==3.1.2
# cffi==1.9.1
click==7.1.2
coverage==4.2
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.3
Flask-Migrate==2.0.2
Flask-Script==2.0.5
Flask-Testing==0.6.1
itsdangerous==0.24
Jinja2==2.11.2
Mako==1.0.6
MarkupSafe==0.23
psycopg2==2.8.5
pycparser==2.17
PyJWT==1.7.1
python-editor==1.0.3
six==1.10.0
SQLAlchemy==1.3.19
Werkzeug==0.16.1
flask-restplus==0.13.0

